server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "django-prometheus (>=2.3.1,<3.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.14.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.3.1,<6.0.0)",
    "googletrans (>=4.0.0-rc1,<5.0.0)" # Ajoutez cette ligne
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"